Register.js functionality is finished
When registering a new user, users are created taking into account their role as student and teacher. It clearly does not allow the user to choose an administrator because this is done with the permission of another administrator from the users section.
And with the backend, encryption is added for the key so that it is not shown to any user.