Login.js functionality is finished
In the login it was modified to have the title and the image of the project.

In the login a role access control was implemented for the functional implementation of restrictions according to permission. There is also a secure authentication system without obvious vulnerabilities and leading to each of the pages with correctly adjusted permissions limiting access by role.

Carroll has access to different menus and functionalities within the application ensuring that a user with a basic role cannot access advanced options, all this working together with the backend and its security functionalities.