Update TeacherHome.js
As on the administrator and student page we have that the teacher has access to different menus and functionalities within the application, in this case the teacher cannot access advanced options as an administrator would but he does have more than a student in this case we have a dashboard that shows how many students and teachers there are, the comparison of current scores and the days playing since the beginning. In the same way, the teacher's explanation is shown so that he knows what he can do. There is the creation of laboratories, equipment, inventories, reservations and can modify the team's reservation. You can view the status of the team, the status of the reservation and you can modify and delete students but not create them.